+++ apps/pie-microsite/jest.config.js
@@ -1,6 +1,13 @@
 module.exports = {
+
+    setupFilesAfterEnv: ['jest-expect-message'],
+
+    testURL: 'http://localhost/',
+
     testEnvironment: 'node',
+
     modulePathIgnorePatterns: [
-        './test/system/'
+        './test/system/',
+        './test/visual/'
     ]
 };