+++ apps/pie-microsite/test/helpers/routes-helper.js
@@ -1,0 +1,39 @@
+const dree = require('dree');
+
+/*
+// Use the 'dree' library to scan the directory of our built pages
+*/
+exports.getNavigationRoutes = () => {
+    const children = dree.scan('./dist/content/pages');
+
+    const expectedRoutes = readChildren(children);
+
+    return expectedRoutes;
+};
+
+/*
+// This function gets called recursively so that subdirectories are included as part of our expected routes.
+*/
+const readChildren = (childDirectories, result = []) => {
+    // Ignore directories that don't have subdirectories / files, as these aren't valid routes
+    if (!childDirectories.children) {
+        return;
+    }
+
+    if (childDirectories.children.length === 1) {
+        if (process.platform === 'win32') {
+            childDirectories.relativePath = childDirectories.relativePath.replaceAll('\\', '/');
+        }
+        result.push(childDirectories.relativePath);
+
+        return;
+    }
+
+    // Recursively call readChildren() so that subdirectories are added to the array.
+    childDirectories.children.forEach(child => {
+        readChildren(child, result);
+    });
+
+    // eslint-disable-next-line consistent-return
+    return result;
+};