+++ apps/pie-microsite/CHANGELOG.md
@@ -3,6 +3,16 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v1.25.0
+------------------------------
+*December 2, 2022*
+
+### Added
+- Added new test related package.json scripts.
+- Add Jest test to compare actual PIE routes to `expected-routes.snapshot.json`.
+- Added WebDriverIO system test to test PIE routes.
+
+
 v1.24.0
 ------------------------------
 *November 29, 2022*